% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data_seurat.R
\name{access_data_seurat}
\alias{access_data_seurat}
\title{Extract cell-level meta data and count matrix from Seurat object}
\usage{
access_data_seurat(seurat_obj, reduction = "umap", output_dir)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{reduction}{Extract x,y coordinates from a type of dimension reduction. If run default \code{harmony_integration()} function, can be one of 'umap' (default) or 'harmony'.}

\item{output_dir}{An output directory}
}
\value{
A list of filtered inputs:
        \itemize{
        \item count: a filtered count matrix
        \item cell_meta: a cell-level metadata
        \item phylo_tree: phylogenetic tree
        }
}
\description{
Extract cell-level meta data and count matrix from Seurat object
}
\examples{
\dontrun{
access_data_seurat(seurat_object,reduction = 'umap')
}

}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
