% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_pseudobulk.R
\name{construct_pseudobulk}
\alias{construct_pseudobulk}
\title{Construct sample-level pseudobulk gene expression matrix}
\usage{
construct_pseudobulk(expr, cell_meta, filter_prop = 0.1)
}
\arguments{
\item{expr}{A raw count gene expression matrix}

\item{cell_meta}{A data frame for cell-level metadata, where each row is a cell. Must contain these columns: barcode, celltype and sample.}

\item{filter_prop}{A number ranges from 0 to 1, to filter low expressed genes across samples (by default: 0.1). Genes with at least this proportion of samples with log2-normalized count greater than 0.01 are retained.}
}
\value{
A sample-level normalized pseudobulk gene expression matrix with genes in the row and samples in the columns
}
\description{
This function takes in raw count gene expression matrix and cell-level metadata as inputs, and generates sample-level normalized pseudobulk gene expression matrix.
}
\examples{
\dontrun{
# default
construct_pseudobulk(expr,cell_meta)
# retain all genes
construct_pseudobulk(expr,cell_meta,filter_prop = 0)
}
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
