% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctprop_cancor.R
\name{ctprop_cancor}
\alias{ctprop_cancor}
\title{Core function for running treecor_ctprop() pipeline.}
\usage{
ctprop_cancor(
  cell_meta,
  sample_meta,
  response_variable,
  method,
  formula = NULL,
  analysis_type = "pearson",
  leaves_info,
  immediate_children,
  component_id = 1,
  num_permutations = 100,
  alternative = "two.sided",
  num_PCs = 2,
  verbose = T
)
}
\arguments{
\item{cell_meta}{A data frame for cell-level metadata, where each row is a cell. Must contain these columns: barcode, celltype and sample.}

\item{sample_meta}{A data frame for sample-level metadata, where each row is a sample. Must contain 'sample' column and additional variables such as covariates or outcome of interest.}

\item{response_variable}{A vector of response variables. Will be used to calculate canonical correlation.}

\item{method}{A character string indicating which approach is used to summarize features. One of 'concat_leaf (default)' or 'concat_immediate_children' or 'aggregate'.}

\item{formula}{An object of class 'formula': a symbolic description of the model to be fitted, adjusting for confounders.}

\item{analysis_type}{'pearson' (default) or 'spearman' correlation or 'regression' for aggregate approach.}

\item{leaves_info}{A data frame that encodes leaf children for each node. Can be extracted from 'hierarchy_list' by running \code{'extract_hrchy_string()'} or \code{'extract_hrchy_seurat()'} function}

\item{immediate_children}{A data frame that contains immediate children for each node. Can be extracted from 'hierarchy_list' by running \code{'extract_hrchy_string()'} or \code{'extract_hrchy_seurat()'} function}

\item{component_id}{An index to extract canonical correlation component (by default: 1).}

\item{num_permutations}{Number of permutations (by default: 1000).}

\item{alternative}{Specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{num_PCs}{Number of PCs}

\item{verbose}{Show progress}
}
\value{
A list of two elements: the first element is a table of summary statistic for each tree node and second element is a list of PC matrices for each tree node.
}
\description{
This function calculates celltype proportion for each sample, performs Principal Component Analysis (PCA), calculates correlation/summary statistic based on top K chosen PCs and obtains permutation based p-value.
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
