% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_cancor.R
\name{expr_cancor}
\alias{expr_cancor}
\title{Compute summary statistic based on gene expression features}
\usage{
expr_cancor(
  pca.ls,
  sample_meta,
  response_variable,
  analysis_type,
  num_permutations,
  alternative,
  component_id = 1,
  ncores
)
}
\arguments{
\item{pca.ls}{A list of cell cluster-specific PCA matrices}

\item{sample_meta}{Sample-level metadata, where each row is a sample. Must contain 'sample' column and additional variables such as covariates or outcome of interest.}

\item{response_variable}{A vector of response variables.}

\item{analysis_type}{Either 'cancor' (canonical correlation) or 'regression' (F-statistic).}

\item{num_permutations}{Number of permutations (by default: 1000).}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{ncores}{Number of cores}

\item{componen_id}{An index to extract canonical correlation component (by default: 1).}
}
\value{
A data frame for summary statistic (canonical correlation or F-stat), p-value and adjusted p-value for each node.
}
\description{
This function takes in a list of cluster-level PCA matrices as input and generates canonical correlation and p-value based on permutation test.
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
