% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hrchy_seurat.R
\name{extract_hrchy_seurat}
\alias{extract_hrchy_seurat}
\title{Extract hierarchy from Seurat object (Data-driven approach)}
\usage{
extract_hrchy_seurat(input, type = "phylo_tree")
}
\arguments{
\item{input}{An input data}

\item{type}{Input data type, must be one of 'phylo_tree' (default) or 'seurat_object'.}
}
\value{
A list of hierarchy information:
        \itemize{
        \item edges: a data frame that documents that parent-children relationship.
        \item layout: a data frame that contains id, xy coordiates, label and a leaf indicator for each node.
        \item immediate_children: a data frame that captures immediate children for each node.
        \item leaves_info: a data frame that summarizes leaf children for each node.
        }
}
\description{
This function takes in either Seurat object or phylogenetic tree and generates hierarchy information.
}
\examples{
# Assume seurat_object is obtained after running treecor_harmony()
extract_hrchy_seurat(seurat_object, type = 'seurat_object')
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
