% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_cellmeta.R
\docType{data}
\name{integrated_cellmeta}
\alias{integrated_cellmeta}
\title{Filtered cell-level metadata after data integration}
\format{
A cell-level metadata dataframe
}
\usage{
data(integrated_cellmeta)
}
\description{
Filtered cell-level metadata after data integration
}
\examples{
data(integrated_cellmeta)
}
\keyword{datasets}
