% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_label.R
\name{modify_label}
\alias{modify_label}
\title{Modify label for tree nodes}
\usage{
modify_label(new_label, hierarchy_list, cell_meta, verbose = T)
}
\arguments{
\item{new_label}{A data frame specifying old and new names for each node. Must contain two columns named as 'old' (corresponds to old celltype name) and 'new' (corresponds to new celltype name).}

\item{hierarchy_list}{A hierarchy list by running \code{extract_hrchy_string()} or \code{extract_hrchy_seurat()} functions, which contains four elements: 'edges', 'layout', 'immediate_children' and 'leaves_info'.}

\item{cell_meta}{Cell-level metadata, where each row is a cell. Must contain these columns: barcode, celltype and sample.}

\item{verbose}{Show progress}
}
\value{
A list that contains modified \code{hierarchy_list} and/or \code{cell_meta}
}
\description{
Rename the node labels
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
