% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_hvgs.R
\name{select_hvgs}
\alias{select_hvgs}
\title{Select highly variable genes (HVGs)}
\usage{
select_hvgs(expr, num_hvgs = NULL)
}
\arguments{
\item{expr}{Sample-level pseudobulk gene expression matrix}

\item{num_hvgs}{Number of highly variable genes extracted from sample-level pseudobulk. Highly variable genes (HVGs) are defined as genes with positive residuals by fitting a gene-specific lowess (\code{loessFit}) function of standard deviation against its mean. The HVGs are ordered by residuals in a descending order. Default is NULL, which includes all genes with positive residuals. Can specify a reasonable positive number (e.g. 1000).}
}
\value{
A vector of highly variable gene names
}
\description{
This function takes in normalized sample-level pseudobulk as input and identifies highly variable genes.
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
