% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecor_celldensityplot.R
\name{treecor_celldensityplot}
\alias{treecor_celldensityplot}
\title{Plot kernel density of cells in a UMAP embedding per group}
\usage{
treecor_celldensityplot(
  cell_meta,
  sample_meta,
  row_variable,
  col_variable = NULL,
  row_combined = T,
  font_size = 15,
  point_size = 0.1,
  point_shape = ".",
  point_alpha = 0.5,
  breaks = seq(0, 1, 0.25)
)
}
\arguments{
\item{cell_meta}{Cell-level metadata. Must contain `sample`,`UMAP_1` and `UMAP_2`columns.}

\item{sample_meta}{Sample-level metadata. Must contain `sample` and variables to be plotted (e.g. row_variable and col_variable).}

\item{row_variable}{Specify variable for stratifying on the row. Must be categorical variable (e.g. `factor` or `character` class).}

\item{col_variable}{Specify variable for  stratifying on the column (optional). Must be categorical variable (e.g. `factor` or `character` class).}

\item{row_combined}{An indicator for combining all subgroups on the rows and visualize the combined plot in the first row. Only works when both `row_variable` and `col_variable` exist.}

\item{font_size}{Font size. Default is 15.}

\item{point_size}{Set point size. Default is 0.1.}

\item{point_shape}{Set point shape. Default is '.'.}

\item{point_alpha}{Set point transparency. Default is 0.5.}

\item{breaks}{A character vectors of breaks for color. Default is c(0,0.25,0.5,0.75,1).}
}
\value{
A kernel density plot in UMAP embedding.
}
\description{
Plot kernel density of cells in a UMAP embedding per group
}
\examples{
treecor_celldensityplot(cell_meta, sample_meta, row_variable = 'severity')
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
