% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecor_ctprop.R
\name{treecor_ctprop}
\alias{treecor_ctprop}
\title{Pipeline for TreeCorTreat using celltype proportion as features}
\usage{
treecor_ctprop(
  hierarchy_list,
  cell_meta,
  sample_meta,
  response_variable,
  method = "aggregate",
  formula = NULL,
  separate = T,
  analysis_type = "pearson",
  num_cancor_components = 1,
  num_permutations = 1000,
  alternative = "two.sided",
  num_PCs = 2,
  verbose = T
)
}
\arguments{
\item{hierarchy_list}{A hierarchy list by running \code{'extract_hrchy_string()'} or \code{'extract_hrchy_seurat()'} function. Contains four elements: 'edges', 'layout', 'immediate_children' and 'leaves_info'.}

\item{cell_meta}{A data frame for cell-level metadata, with cells on the row. Must contain these columns: 'barcode', 'celltype' and 'sample'.}

\item{sample_meta}{A data frame for sample-level metadata, where each row is a sample. Must contain 'sample' column and additional variables to be used in the analysis, such as covariates or outcomes of interest.}

\item{response_variable}{A vector of response variables. Extract these columns from 'sample_meta'.}

\item{method}{A character string indicating which approach is used to summarize features. One of 'aggregate' (default) or 'concat_leaf' or 'concat_immediate_children'.}

\item{formula}{An object of class 'formula': a symbolic description of the model to be fitted, adjusting for confounders.}

\item{separate}{A TRUE (default) or FALSE indicator, specifying how to evaluate multivariate outcomes.
\itemize{
\item TRUE: evaluate multivariate phenotype separately (it is equivalent to run this pipeline for each univariate phenotype).
\item FALSE: evaluate multivariate phenotype jointly.
}}

\item{analysis_type}{Specify a summary statistic for analysis, must be one of
\itemize{
\item pearson (default): Pearson correlation (only for \code{method = 'aggregate'})
\item spearman: Spearman correlation (only for \code{method = 'aggregate'})
\item cancor: Canonical correlation
\item regression: Fit a linear regression with cell type proportion feature vectors as dependent variable and phenotype(s) as explanatory variables (full model). Compare between full model (with phenotypes in explanatory variables) and reduced model (without phenotypes).
}}

\item{num_cancor_components}{Number of canonical components to be extracted. Only works for multivariate phenotypes when specifying \code{'separate = F'} and \code{'analysis_type = cancor'}.}

\item{num_permutations}{Number of permutations (by default: 1000).}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{num_PCs}{Number of PCs}

\item{verbose}{Show progress}
}
\value{
A list of two elements: the first element is a table of summary statistic for each tree node and second element is a list of PC matrices for each tree node.
}
\description{
Pipeline for TreeCorTreat using celltype proportion as features
}
\examples{
# default setting
result <- treecor_ctprop(hierarchy_list, cell_meta, sample_meta, response_variable = 'severity')
# extract a table of summary statistic for each cell cluster
result[[1]] # or result$canonical_corr
# extract a list of PC matrices for each cell cluster
result[[2]] # or result$pc.ls
# extract PC matrix for celltype 'T'
result$pc.ls[['T']]
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
