% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecor_degheatmap.R
\name{treecor_degheatmap}
\alias{treecor_degheatmap}
\title{Plot heatmap for top n differentially expressed genes}
\usage{
treecor_degheatmap(
  sample_meta,
  pseudobulk,
  deg_result,
  top_n = 10,
  deg_logFC = "both",
  scale = "row",
  cluster_rows = F,
  cluster_cols = F,
  show_colnames = F,
  show_rownames = T,
  annotation_col = NA,
  annotation_colors = NA,
  color = colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(100),
  breaks = NA,
  fontsize = 12
)
}
\arguments{
\item{sample_meta}{A data frame for sample-level metadata, where each row is a sample. Must contain 'sample' column and additional variables such as covariates or outcome of interest.}

\item{pseudobulk}{A sample-level pseudobulk, which can be obtained by running `limma_expr` pipeline and can be extracted via `result$pseudobulk.ls[[cell_type_name]]`.}

\item{deg_result}{Differential expression output obtained by running limma_expr` pipeline` and can be extracted via `result$dge.ls[[phenotype]][[cell_type_name]]`.}

\item{top_n}{A number to extract top n differentially expressed genes. If \code{deg_logFC = 'both'}, then top n DEGs with positive logFC will be combined with top n DEGs with negative logFC, resulting in a total of 2n genes.}

\item{deg_logFC}{Filter `deg_result` by logFC. One of the following:
#' \itemize{
\item positive: limit to genes with logFC>0.
\item negative: limit to genes with logFC<0.
\item both (default): keep genes with both positive and negative logFC.
}}

\item{scale}{A character indicates if the values should be centered and scaled in either the `row` direction or the `column` direction, or `none`.}

\item{cluster_rows}{A True of False indicator to determine if rows should be clustered.}

\item{cluster_cols}{A True of False indicator to determine if column names should be displayed.}

\item{annotation_col}{A vector specifying column names to be annotated in the sidebars of a heatmap.}

\item{annotation_colors}{List for manually specifying colors for annotation_col.}

\item{color}{A vector of colors used in heatmap.}

\item{breaks}{A sequence of numbers specifying range of values in the matrix. Must be an element longer than color vector. See more in `?pheatmap`.}

\item{fontsize}{Font size.}

\item{show_rows}{A True of False indicator to determine if row names should be displayed.}
}
\value{
Heatmap with DEGs on the row and samples on the column.
}
\description{
Plot heatmap for top n differentially expressed genes
}
\examples{
# Assume obtain pseudobulk of celltype 'T' from treecor_deg()
result <- treecor_deg(expr,hierarchy_list, cell_meta, sample_meta, response_variable = 'severity')
pseudobulk <- result$pseudobulk.ls[['T']]
deg_result <- result$dge.ls$severity[['T']]
# Top 10 DEGs with logFC>0
treecor_degheatmap(sample_meta,pseudobulk, deg_result, top_n = 10, deg_logFC = 'positive')
# Top 10 DEGs with logFC<0
treecor_degheatmap(sample_meta,pseudobulk, deg_result, top_n = 10, deg_logFC = 'negative')
# Combined the above two options: top 10 positively-associated DEGs and top 10 negatively-associated DEGs
treecor_degheatmap(sample_meta,pseudobulk, deg_result, top_n = 10, deg_logFC = 'both')
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
