% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecor_samplepcaplot.R
\name{treecor_samplepcaplot}
\alias{treecor_samplepcaplot}
\title{Sample PCA plot}
\usage{
treecor_samplepcaplot(
  sample_meta,
  pca_matrix,
  response_variable,
  analysis_type = "cancor",
  num_permutations = 1000,
  alternative = "two.sided",
  row_variable = NULL,
  col_variable = NULL,
  font_size = 15,
  point_size = 1,
  point_shape = 19,
  point_alpha = 1,
  line_type = "dashed",
  line_color = "black"
)
}
\arguments{
\item{sample_meta}{Sample-level metadata. Must contain \code{sample} and variables to be plotted (e.g. \code{row_variable} and \code{col_variable}).}

\item{pca_matrix}{PCA matrix with sample names on the rows and PC coordinates on the columns (e.g. PC1 and PC2).}

\item{response_variable}{A univariate phenotype. If multivariate phenotype is provided, either convert them to a univariate combined phenotype (e.g. using weight or PCA) or call this function separately for each outcome.}

\item{analysis_type}{Specify summary statistic, can be one of 'cancor' (default) or 'regression'.}

\item{num_permutations}{Number of permutations. Default is 1000.}

\item{alternative}{Specify alternative hypothesis. Must be one of 'two.sided' (default) or 'less' or 'greater'.}

\item{row_variable}{Specify variable for row. Must be categorical variable (e.g. `factor` or `character` class).}

\item{col_variable}{Specify variable for column (optional). Must be categorical variable (e.g. `factor` or `character` class).}

\item{font_size}{Font size. Default is 15.}

\item{point_size}{Set point size. Default is 0.1.}

\item{point_shape}{Set point shape. Default is 19.}

\item{point_alpha}{Set point transparency. Default is 1.}

\item{line_type}{Set line type as 'dashed' (default) or 'solid' or others.}

\item{line_color}{Set optimal axis line color.}
}
\value{
Sample pca plot.
}
\description{
Sample PCA plot
}
\examples{
treecor_samplepcaplot(sample_meta, pca_matrix, response_variable = 'severity', row_variable = 'severity')
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
